requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
fonttools==4.53.1
psycopg2==2.9.10
flasgger==0.9.7.1
Werkzeug==3.0.1
pytest==8.3.4
PyJWT==2.10.1